requirements.txt

requests
python-dotenv
pandas
tqdm
streamlit
matplotlib
numpy
